backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
pyjwt
psycopg2-binary
python-dotenv
pytest
httpx
requests
passlib[bcrypt]
python-multipart